Core enemybase enhanced
