+++ Assets/Scripts/Core/Enemies/EnemyBase.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections;
+﻿using DG.Tweening;
+using System.Collections;
 using UnityEngine;
 
 namespace Assets.Scripts.Core.Enemies
@@ -17,12 +18,41 @@ namespace Assets.Scripts.Core.Enemies
         [SerializeField] private float _flashSpeed = 1f;
         [SerializeField] private ParticleSystem _damageVFX;
 
-        protected virtual void PlayDamageVfx()
+        private Tween _currentColorTween;
+
+        protected virtual void PlayDamageFeedback()
+        {
+            PlayParticleSystem();
+            FlashShader();
+        }
+        private void PlayParticleSystem()
         {
             if (_damageVFX != null)
             {
                 _damageVFX.Play();
             }
+        }
+        private void FlashShader()
+        {
+            if (!_currentColorTween.IsActive())
+            {
+                _currentColorTween = _enemyMeshRenderer.material.DOColor(_damageFeedbackColor, "_EmissionColor", _flashSpeed).SetLoops(2, LoopType.Yoyo);
+                _damageVFX.Play();
+            }
+        }
+
+        protected virtual void Update()
+        {
+            IsAlive = CheckIfHasHealthPoints();
+        }
+        private bool CheckIfHasHealthPoints()
+        {
+            if (_healthPoints > 0)
+            {
+                return true;
+            }
+
+            return false;
         }
     }
 }
