--- Assets/Scripts/Enemies/EnemyBase.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1b567fff49e8e4b4e8824d76a6a7a9cb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
